+++ src/components/NextMeetup.tsx
@@ -2,7 +2,6 @@ import axios from "axios";
 import { useEffect, useState } from "react";
 import { MeetupEvent, MeetupResponse } from "../types/app";
 import dayjs from "dayjs";
-import Marquee from "react-fast-marquee";
 import { PizzaDay } from "./PizzaDay";
 
 export const NextMeetup = () => {
@@ -19,24 +18,18 @@ export const NextMeetup = () => {
   };
 
   const showPizzaDay = () => {
-    const today = dayjs().format("YYYY-MM-DD");
+    // const today = dayjs().format("YYYY-MM-DD");
 
-    // show pizza day if less than 60 days before 22 may
-    const year = dayjs().year();
-    const pizzaDay = dayjs(`${year}-05-22`)
-      .subtract(60, "day")
-      .format("YYYY-MM-DD");
-    return today < pizzaDay;
+    // // show pizza day if less than 60 days before 22 may
+    // const year = dayjs().year();
+    // const pizzaDay = dayjs(`${year}-05-22`);
+
+    // const delta = pizzaDay.diff(today, "day");
+
+    // return delta < 300;
+
+    return true;
   };
-
-  const pastEventImageUrls = [
-    "https://bafybeibthopyr5u3txthpwblottoxnp3hiddrvdafz2yk2m4qgl5prs6ru.ipfs.dweb.link/",
-    "https://bafybeibhvh6qwznkwcqc5rwgljrvvbs733youhn53d2h3y4mwxc33mogem.ipfs.dweb.link/",
-    "https://bafybeiaqwbqi76zca3so72xksop7flljm6zcqgdod7fg5o5b5am5tjzrje.ipfs.dweb.link/",
-    "https://bafybeidxjf5oc3ho6pgqvmtofoye7pqhq7eeohq2af2jmwhk6kmrpzinva.ipfs.dweb.link/",
-    "https://sgp1.digitaloceanspaces.com/cms.dltx.io/7adea14177b41930ae385aa65425dce3.jpg",
-    "https://bafybeidgem5sjv4frkoi622gfbmrfwgreofkc7vekyjjvv4ieris5anh74.ipfs.dweb.link"
-  ];
 
   useEffect(() => {
     axios
@@ -109,9 +102,9 @@ export const NextMeetup = () => {
         </div>
       </div>
 
-      {showPizzaDay() && <PizzaDay/>}
+      {showPizzaDay() && <PizzaDay />}
 
-      <div className="flex flex-col text-center">
+      {/* <div className="flex flex-col text-center">
         <p className="mt-48 md:text-4xl text-xl font-hand">
           Checkout our other events @
           <a
@@ -123,22 +116,7 @@ export const NextMeetup = () => {
             meetup.com/bitcoinbrisbane
           </a>
         </p>
-      </div>
-
-      {/* Carousel of past event images */}
-      {pastEventImageUrls.length > 0 && (
-        <div className="flex flex-col text-center">
-          <Marquee speed={82}>
-            {pastEventImageUrls.map(imageUrl => (
-              <img
-                src={imageUrl}
-                alt="past event image"
-                className="md:h-[20rem] object-cover rounded-xl m-2 h-[8rem]"
-              />
-            ))}
-          </Marquee>
-        </div>
-      )}
+      </div> */}
     </section>
   );
 };
